Move chat message tile props to common
