+++ packages/common/src/store/pages/chat/types.ts
@@ -13,3 +13,10 @@ export enum ChatPermissionAction {
   /** User is signed out and needs to sign in */
   SIGN_UP
 }
+
+export type ChatMessageTileProps = {
+  link: string
+  onEmpty?: () => void
+  onSuccess?: () => void
+  className?: string
+}
